requirements.txt
pandas
psycopy2
psycopg2-binary
requests
scikit-learn
matplotlib
seaborn
numpy
scipy
python-dateutil